后端：做了细节处理，添加静态资源配置，完成了若干接口。

+++ src/main/java/com/lms/dao/DocListDao.java
@@ -11,7 +11,7 @@ import java.util.List;
 
 public interface DocListDao extends JpaRepository<DocList, Integer> {
     DocList findByName(String name);
-    DocList findByDlid(Long dlid);
+    DocList findByDlid(Integer dlid);
     Page<DocList> findAll(Pageable pageable);
     List<DocList> findByUserEntity(UserEntity uiid);
     List<DocList> findAllByOrderByCreatetimeDesc();
